+++ packages/kryo/src/lib/builtins/any.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/any
- */
-
-import { AnyType } from "../any.js";
-
-export const $Any: AnyType = new AnyType();