+++ packages/kryo/src/lib/builtins/bytes.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/bytes
- */
-
-import { BytesType } from "../bytes.js";
-
-export const $Bytes: BytesType = new BytesType({maxLength: Infinity});