+++ packages/kryo/src/lib/builtins/uint8.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/uint8
- */
-
-import { IntegerType } from "../integer.js";
-
-export const $Uint8: IntegerType = new IntegerType({min: 0, max: 255});